+++ integration_tests/wordcount/test_new_data.py
@@ -8,6 +8,7 @@ import pathlib
 import pytest
 
 from .base import (
+    AZURE_STORAGE_NAME,
     FS_STORAGE_NAME,
     INPUT_PERSISTENCE_MODE_NAME,
     OPERATOR_PERSISTENCE_MODE_NAME,
@@ -19,7 +20,9 @@ from .base import (
 
 
 @pytest.mark.parametrize("n_threads,n_processes", [(1, 1), (4, 1), (1, 4), (2, 2)])
-@pytest.mark.parametrize("pstorage_type", [S3_STORAGE_NAME, FS_STORAGE_NAME])
+@pytest.mark.parametrize(
+    "pstorage_type", [S3_STORAGE_NAME, FS_STORAGE_NAME, AZURE_STORAGE_NAME]
+)
 @pytest.mark.parametrize(
     "persistence_mode", [INPUT_PERSISTENCE_MODE_NAME, OPERATOR_PERSISTENCE_MODE_NAME]
 )