+++ src/persistence/backends/s3.rs
@@ -1,13 +1,5 @@
 // Copyright © 2024 Pathway
 
-use log::error;
-use std::mem::take;
-use std::sync::mpsc;
-use std::sync::mpsc::Sender;
-use std::thread;
-
-use futures::channel::oneshot;
-use futures::channel::oneshot::Sender as OneShotSender;
 use s3::bucket::Bucket as S3Bucket;
 
 use crate::deepcopy::DeepCopy;
@@ -15,27 +7,16 @@ use crate::persistence::backends::PersistenceBackend;
 use crate::persistence::Error;
 use crate::retry::{execute_with_retries, RetryConfig};
 
-use super::BackendPutFuture;
+use super::{BackendPutFuture, BackgroundObjectUploader};
 
 const MAX_S3_RETRIES: usize = 2;
-
-#[derive(Debug)]
-enum S3UploaderEvent {
-    UploadObject {
-        key: String,
-        value: Vec<u8>,
-        result_sender: OneShotSender<Result<(), Error>>,
-    },
-    Finish,
-}
 
 #[derive(Debug)]
 #[allow(clippy::module_name_repetitions)]
 pub struct S3KVStorage {
     bucket: S3Bucket,
     root_path: String,
-    upload_event_sender: Sender<S3UploaderEvent>,
-    uploader_thread: Option<std::thread::JoinHandle<()>>,
+    background_uploader: BackgroundObjectUploader,
 }
 
 impl S3KVStorage {
@@ -45,55 +26,25 @@ impl S3KVStorage {
             root_path_prepared += "/";
         }
 
-        let (upload_event_sender, upload_event_receiver) = mpsc::channel();
         let uploader_bucket = bucket.deep_copy();
-        let uploader_thread = thread::Builder::new()
-            .name("pathway:s3_snapshot-bg-writer".to_string())
-            .spawn(move || {
-                loop {
-                    let event = upload_event_receiver.recv().expect("unexpected termination for s3 objects sender");
-                    match event {
-                        S3UploaderEvent::UploadObject { key, value, result_sender } => {
-                            let upload_result = execute_with_retries(
-                                || uploader_bucket.put_object(&key, &value),
-                                RetryConfig::default(),
-                                MAX_S3_RETRIES,
-                            ).map_err(Error::S3).map(|_| ());
-                            let send_result = result_sender.send(upload_result);
-                            if let Err(unsent_flush_result) = send_result {
-                                error!("The receiver no longer waits for the result of this save: {unsent_flush_result:?}");
-                            }
-                        },
-                        S3UploaderEvent::Finish => break,
-                    };
-                }
-            })
-            .expect("s3 uploader failed");
+        let upload_object = move |key: String, value: Vec<u8>| {
+            let _ = execute_with_retries(
+                || uploader_bucket.put_object(&key, &value),
+                RetryConfig::default(),
+                MAX_S3_RETRIES,
+            )?;
+            Ok(())
+        };
 
         Self {
             bucket,
-            upload_event_sender,
-            uploader_thread: Some(uploader_thread),
+            background_uploader: BackgroundObjectUploader::new(upload_object),
             root_path: root_path_prepared,
         }
     }
 
     fn full_key_path(&self, key: &str) -> String {
         self.root_path.clone() + key
-    }
-}
-
-impl Drop for S3KVStorage {
-    fn drop(&mut self) {
-        self.upload_event_sender
-            .send(S3UploaderEvent::Finish)
-            .expect("failed to submit the graceful shutdown event");
-        if let Some(uploader_thread) = take(&mut self.uploader_thread) {
-            if let Err(e) = uploader_thread.join() {
-                // there is no formatter for std::any::Any
-                error!("Failed to join s3 snapshot uploader thread: {e:?}");
-            }
-        }
     }
 }
 
@@ -109,11 +60,7 @@ impl PersistenceBackend for S3KVStorage {
                 let key: &str = &object.key;
                 assert!(key.len() > self.root_path.len());
                 let prepared_key = key[prefix_len..].to_string();
-                if !prepared_key.contains('/') {
-                    // Similarly to the filesystem backend,
-                    // we take only files in the imaginary folder
-                    keys.push(prepared_key);
-                }
+                keys.push(prepared_key);
             }
         }
 
@@ -131,15 +78,8 @@ impl PersistenceBackend for S3KVStorage {
     }
 
     fn put_value(&mut self, key: &str, value: Vec<u8>) -> BackendPutFuture {
-        let (sender, receiver) = oneshot::channel();
-        self.upload_event_sender
-            .send(S3UploaderEvent::UploadObject {
-                key: self.full_key_path(key),
-                value,
-                result_sender: sender,
-            })
-            .expect("chunk queue submission should not fail");
-        receiver
+        self.background_uploader
+            .upload_object(self.full_key_path(key), value)
     }
 
     fn remove_key(&mut self, key: &str) -> Result<(), Error> {